AudioCache

+++ TagTool/Commands/Porting/PortTagCommand.Sound.cs
@@ -12,6 +12,8 @@ using System.Linq;
 using TagTool.Tags;
 using TagTool.Audio;
 using System.Threading;
+using System.Security.Cryptography;
+using System.Text;
 
 namespace TagTool.Commands.Porting
 {
@@ -95,7 +97,7 @@ namespace TagTool.Commands.Porting
             return header;
         }
 
-        public static string ConvertSoundPermutation(byte[] buffer, int index, int count, int fileSize, byte channelCount, SampleRate sampleRate, bool loop)
+        public static string ConvertSoundPermutation(byte[] buffer, int index, int count, int fileSize, byte channelCount, SampleRate sampleRate, bool loop, bool use_cache, string permutation_mp3_cachename)
         {
             Directory.CreateDirectory(@"Temp");
 
@@ -105,11 +107,13 @@ namespace TagTool.Commands.Porting
                 return null;
             }
 
-            string tempXMA = @"Temp\permutation.xma";
-            string tempWAV = @"Temp\permutation.wav";
-            string fixedWAV = @"Temp\permutationTruncated.wav";
-            string loopMP3 = @"Temp\permutationTruncated.mp3";
-            string tempMP3 = @"Temp\permutation.mp3";
+            string audiofile = permutation_mp3_cachename ?? "Temp\\permutation";
+
+            string tempXMA = $"{audiofile}.xma";
+            string tempWAV = $"{audiofile}.wav";
+            string fixedWAV = $"{audiofile}Truncated.wav";
+            string loopMP3 = $"{audiofile}Truncated.mp3";
+            string tempMP3 = $"{audiofile}.mp3";
 
             //If the files are still present, somehow, before the conversion happens, it will stall because ffmpeg doesn't override existing sounds.
 
@@ -274,12 +278,19 @@ namespace TagTool.Commands.Porting
             }
             finally
             {
-                if (File.Exists(tempXMA))
-                    File.Delete(tempXMA);
-                if (File.Exists(tempWAV))
-                    File.Delete(tempWAV);
-                if (File.Exists(fixedWAV))
-                    File.Delete(fixedWAV);
+                if(!use_cache)
+                {
+                    if (File.Exists(tempXMA))
+                        File.Delete(tempXMA);
+                    if (File.Exists(tempWAV))
+                        File.Delete(tempWAV);
+                    if (File.Exists(fixedWAV))
+                        File.Delete(fixedWAV);
+                    if (File.Exists(loopMP3))
+                        File.Delete(loopMP3);
+                    if (File.Exists(tempMP3))
+                        File.Delete(tempMP3);
+                }
             }
             return tempMP3;
         }
@@ -294,7 +305,25 @@ namespace TagTool.Commands.Porting
             return 10.0f * (float)Math.Log10(ratio);
         }
 
-        private Sound ConvertSound(Stream cacheStream, Dictionary<ResourceLocation, Stream> resourceStreams, Sound sound)
+        static string ComputeSha256Hash(string rawData)
+        {
+            // Create a SHA256   
+            using (SHA256 sha256Hash = SHA256.Create())
+            {
+                // ComputeHash - returns byte array  
+                byte[] bytes = sha256Hash.ComputeHash(System.Text.Encoding.UTF8.GetBytes(rawData));
+
+                // Convert byte array to a string   
+                StringBuilder builder = new StringBuilder();
+                for (int i = 0; i < bytes.Length; i++)
+                {
+                    builder.Append(bytes[i].ToString("x2"));
+                }
+                return builder.ToString();
+            }
+        }
+
+        private Sound ConvertSound(Stream cacheStream, Dictionary<ResourceLocation, Stream> resourceStreams, Sound sound, string blamTag_Name)
         {
             if (BlamSoundGestalt == null)
                 BlamSoundGestalt = PortingContextFactory.LoadSoundGestalt(CacheContext, ref BlamCache);
@@ -452,6 +481,8 @@ namespace TagTool.Commands.Porting
                 //
 
                 permutationIndex = firstPermutationIndex;
+                bool use_cache = Sounds.UseAudioCacheCommand.AudioCacheDirectory != null;
+                var base_permutation_mp3_cachename = blamTag_Name != null ? ComputeSha256Hash(blamTag_Name) : null;
 
                 for (i = 0; i < permutationCount; i++)
                 {
@@ -484,7 +515,11 @@ namespace TagTool.Commands.Porting
                         loop = true;
                     }
 
-                    var permutationMP3 = ConvertSoundPermutation(resourceData, begin, count, count + 52, (byte)channelCount, sound.SampleRate, loop);
+                    string permutation_mp3_cachename = use_cache ? $"{base_permutation_mp3_cachename}_{i}" : null;
+                    if(permutation_mp3_cachename == null || !File.Exists($"{permutation_mp3_cachename}.mp3"))
+                    {
+                        permutation_mp3_cachename = ConvertSoundPermutation(resourceData, begin, count, count + 52, (byte)channelCount, sound.SampleRate, loop, use_cache, permutation_mp3_cachename);
+                    }
 
                     uint permutationChunkSize = 0;
 
@@ -492,7 +527,7 @@ namespace TagTool.Commands.Porting
                     // Copy the permutation mp3 to the overall mp3
                     //
 
-                    byte[] permBuffer = File.ReadAllBytes(permutationMP3);
+                    byte[] permBuffer = File.ReadAllBytes(permutation_mp3_cachename);
                     try
                     {
                         using (Stream output = new FileStream(soundMP3, FileMode.Append, FileAccess.Write, FileShare.None))
